add .gitattributes and .gitignore (#2)
The .gitignore file tells Git to ignore files matching the description within. Usually repositories hold only the source file. So in this case the compiled Java .class files shoud be excluded.

The .gitattributes file carries instructions for Git to handle certain file types. I added this because the End Of Line in *nix systems are LF and CRLF in windows. This may cause problems when files are being edited concurrently by users in Windows that have not set EOL to LF (if EOL is changed from LF to CRLF, git may think every line is changed).